# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/token": {
      "post": {
        "tags": ["authenticaton"],
        "summary": "Login",
        "operationId": "login_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/refresh-token": {
      "post": {
        "tags": ["authenticaton"],
        "summary": "Refresh Token",
        "operationId": "refresh_token_refresh_token_post",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": ["authenticaton"],
        "summary": "Read Users Me",
        "operationId": "read_users_me_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/users/proposals/": {
      "post": {
        "tags": ["user"],
        "summary": "Add Proposal",
        "operationId": "add_proposal_users_proposals__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/proposals/{proposal_id}": {
      "get": {
        "tags": ["user"],
        "summary": "Read Proposal",
        "operationId": "read_proposal_users_proposals__proposal_id__get",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Proposal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/projects/": {
      "get": {
        "tags": ["user"],
        "summary": "Read Projects",
        "operationId": "read_projects_users_projects__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "title": "Response Read Projects Users Projects  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/reports-by-project/{project_id}": {
      "get": {
        "tags": ["user"],
        "summary": "Read Reports",
        "operationId": "read_reports_users_reports_by_project__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportResponse"
                  },
                  "title": "Response Read Reports Users Reports By Project  Project Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/reports/": {
      "post": {
        "tags": ["user"],
        "summary": "Add Report",
        "operationId": "add_report_users_reports__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/reports/{report_id}": {
      "get": {
        "tags": ["user"],
        "summary": "Read Report",
        "operationId": "read_report_users_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_login_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": ["username", "password"],
        "title": "Body_login_token_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "proposal_id": {
            "type": "integer",
            "title": "Proposal Id"
          },
          "user_supervisor_id": {
            "type": "integer",
            "title": "User Supervisor Id"
          },
          "user_discoverer_id": {
            "type": "integer",
            "title": "User Discoverer Id"
          },
          "user_master_id": {
            "type": "integer",
            "title": "User Master Id"
          },
          "user_broker_id": {
            "type": "integer",
            "title": "User Broker Id"
          },
          "user_user_id": {
            "type": "integer",
            "title": "User User Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "proposal_id",
          "user_supervisor_id",
          "user_discoverer_id",
          "user_master_id",
          "user_broker_id",
          "user_user_id"
        ],
        "title": "ProjectResponse"
      },
      "ProposalRequest": {
        "properties": {
          "info": {
            "type": "string",
            "title": "Info"
          },
          "RFP_id": {
            "type": "integer",
            "title": "Rfp Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "state": {
            "type": "integer",
            "title": "State"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Id"
          }
        },
        "type": "object",
        "required": ["info", "RFP_id", "user_id", "state", "comment"],
        "title": "ProposalRequest"
      },
      "ProposalResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "info": {
            "type": "string",
            "title": "Info"
          },
          "RFP_id": {
            "type": "integer",
            "title": "Rfp Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "state": {
            "type": "integer",
            "title": "State"
          },
          "file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Id"
          }
        },
        "type": "object",
        "required": ["id", "info", "RFP_id", "user_id", "comment", "state"],
        "title": "ProposalResponse"
      },
...
      "ReportRequest": {
        "properties": {
          "info": {
            "type": "string",
            "title": "Info"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "state": {
            "type": "integer",
            "title": "State"
          }
        },
        "type": "object",
        "required": ["info", "project_id", "comment", "state"],
        "title": "ReportRequest"
      },
      "ReportResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "info": {
            "type": "string",
            "title": "Info"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "state": {
            "type": "integer",
            "title": "State"
          }
        },
        "type": "object",
        "required": ["id", "info", "project_id", "comment", "state"],
        "title": "ReportResponse"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": ["access_token", "token_type", "refresh_token"],
        "title": "Token"
      },
      "UserInfoResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "fname": {
            "type": "string",
            "title": "Fname"
          },
          "lname": {
            "type": "string",
            "title": "Lname"
          },
          "father_name": {
            "type": "string",
            "title": "Father Name"
          },
          "birth": {
            "type": "string",
            "format": "date",
            "title": "Birth"
          },
          "resume_file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume File Id"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "user_type": {
            "$ref": "#/components/schemas/UserTypeInfo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "fname",
          "lname",
          "father_name",
          "birth",
          "active",
          "user_type"
        ],
        "title": "UserInfoResponse"
      },
      "UserTypeInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "info": {
            "type": "string",
            "title": "Info"
          }
        },
        "type": "object",
        "required": ["id", "info"],
        "title": "UserTypeInfo"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}